Use non-greedy regex
The `orderless-flex' matching style will match input "foobr-mode" to
both "foobar-mode" and "foo-bar-baz-quux-mode", but chances are that
the user means "foobar-mode". The same is for `orderless-prefix' and
`orderless-initialism'.

By using non-greedy regexes, the shorter matches will be listed first,
which is likely what users want, rather than the longest match.