Implement ^literal optimization discussed in #79
If orderless finds that some component of the pattern compiles to a
regexp of the form ^literal it will now take the first such regexp and
add the literal to the prefix which which it calls all-completions.

That's a slight lie: it actually looks for regexps of the form
\(?:^literal\) since that is what the pattern compiler actually
produces.

With this change people can now add a style dispatcher that adds a ^
to the first component to use with certain completion-at-point
functions that can really take advantage of knowing a prefix to return
fewer completions. Some even (incorrectly) refuse to return all
possible completions for an empty string!

See the discussion in #79 and the issues linked therein.