Remove strict initialism code by @noctuid
The function `orderless--separated-by` has also been added by @noctuid. However
with the removal of the strict initialism code, the contributions by @noctuid
are pushed below the 15 lines limit. The function `orderless--separated-by` has
only 11 lines.

(cl-defun orderless--separated-by (sep rxs &optional (before "") (after ""))
  "Return a regexp to match the rx-regexps RXS with SEP in between.
If BEFORE is specified, add it to the beginning of the rx sequence. If AFTER is
specified, add it to the end of the rx sequence."
  (rx-to-string
   `(seq
     ,before
     ,@(cl-loop for (sexp . more) on rxs
                collect `(group ,sexp)
                when more collect `,sep)
     ,after)))