Remove orderless-skip-highlighting
This variable was added to support Selectrum, which has been deprecated now.
Vertico and Corfu use a different technique to implement deferred highlighting
for the visible candidates only.

There are no other users of this variable. The problem of deferred highlighting
must be solved more generally by improving the completion styles API itself in
minibuffer.el. It is therefore better to not provide this variable here.