Work around inconsistency of `completion-table-with-context'
The function `completion-table-with-context' modifies the predicate such that
the string arguments are prefixed. This behavior is inconsistent with the
`completion-file-name-table', where the strings returned by `all-completions'
are the same as the strings passed to the predicate. The behavior of
`all-completions' seems more intuitive at first sight. This could be an
unintended bug or oversight in `completion-table-with-context'.

In `orderless-try-completion' we work around the inconsistency by checking if a
prefix is already present in the string passed to the predicate. If not, the
prefix is added.

Fix #148