Add `orderless-expand-substring` for literal substring expansion
If the customization option `orderless-expand-substring` is set to
t, `orderless-try-completion` first tries to treat the input as
literal substring and tries to expand it. If literal substring
expansion fails, Orderless falls back to its regular behavior, and
checks if there exists a single unique match to expand to. The
option `orderless-expand-substring` is effective in the default
completion UI and Corfu when pressing TAB, but does not affect
Vertico. It makes the transition to Orderless a little friendlier
for users accustomed to the Emacs built-in `basic` and `substring`
completion styles.

As an alternative to `orderless-expand-substring` one could define
an alternative completion style `tab` as follows.

    (add-to-list 'completion-styles-alist
                 '(tab completion-substring-try-completion ignore
                 "Completion style which provides only TAB completion."))

I've used this `tab` configuration style for a long time in my
personal configuration, but the goal is to improve out of the box
behavior of Orderless and related packages. The following two
configurations are equivalent:

    ;; NEW: Use orderless-expand-substring
    (setq completion-styles '(orderless basic)
          orderless-expand-substring t)
    ;; OLD: Use the tab style from above.
    (setq completion-styles '(tab orderless basic)
          orderless-expand-substring nil)