orderless-expand-prefix: Support the values `prefix`, `substring` and `nil`
Using `completion-substring-try-completion` can be costly in comparison to
prefix completion only via `try-completion`, in particular in the context of
timer-based auto completion. As a compromise, default to `prefix` completion
only. Performance should not be affected, while expansion is still provided.
Depending on preference the expansion setting can be escalated to `substring` or
even disabled, in case the restrictive Orderless (non-)expansion behavior is
preferred.